use reboot command for mini OG

+++ sprig/scripts/save_poweroff.sh
@@ -21,4 +21,9 @@ for emulator in $EMU_LIST; do
 done
 
 sync
-poweroff
+
+if is_mini_og; then
+    reboot          # OG Mini hangs if you use the poweroff command
+else
+    poweroff
+fi
